added tests for loader
