+++ tests/unit/all.ts
@@ -3,3 +3,4 @@ import './plugins/CoreLoadPlugin';
 import './plugins/I18nPlugin';
 import './plugins/InjectModulesPlugin';
 import './plugins/util';
+import './loaders/css-module-decorator-loader';
