+++ tests/unit/loaders/css-module-decorator-loader.ts
@@ -1,0 +1,32 @@
+import { describe, it } from 'intern!bdd';
+import * as assert from 'intern/chai!assert';
+// import * as path from 'path';
+// import * as sinon from 'sinon';
+import loader from '../../../src/loaders/css-module-decorator-loader/loader';
+
+describe('css-module-decorator-loader', () => {
+	it('should not effect content without local exports', () => {
+		const content = `exports = 'abc'
+		exports.push(['a', 'b'])`;
+
+		const result = loader(content);
+		assert.equal(result, content);
+	});
+
+	it('should wrap local exports with decorator', () => {
+		const content = `exports.locals = { "hello": "world" };`;
+
+		const result = loader.bind({ resourcePath: 'testFile.css' })(content);
+		assert.equal(result, 'exports.locals = {"default":{"classes":{"hello":"world"},"path":"testFile"}};');
+	});
+
+	it('should wrap multi line local exports with decorator', () => {
+		const content = `exports.locals = {
+			"hello": "world",
+			"foo": "bar"
+		};`;
+
+		const result = loader.bind({ resourcePath: 'testFile.css' })(content);
+		assert.equal(result, 'exports.locals = {"default":{"classes":{"hello":"world","foo":"bar"},"path":"testFile"}};');
+	});
+});
